+++ PlotsBase/src/utils.jl
@@ -5,9 +5,6 @@ else
     tempdir()
 end
 
-treats_y_as_x(seriestype) =
-    seriestype in (:vline, :vspan, :histogram, :barhist, :stephist, :scatterhist)
-
 function replace_image_with_heatmap(z::AbstractMatrix{<:Colorant})
     n, m = size(z)
     colors = palette(vec(z))
@@ -395,17 +392,6 @@ function Commons.preprocess_attributes!(plotattributes::AKW)
                 process_axis_arg!(plotattributes, arg, letter)
             end
         end
-    end
-
-    # vline and others accesses the y argument but actually maps it to the x axis.
-    # Hence, we have to take care of formatters
-    if treats_y_as_x(get(plotattributes, :seriestype, :path))
-        xformatter = get(plotattributes, :xformatter, :auto)
-        yformatter = get(plotattributes, :yformatter, :auto)
-        yformatter ≢ :auto && (plotattributes[:xformatter] = yformatter)
-        xformatter ≡ :auto &&
-            haskey(plotattributes, :yformatter) &&
-            pop!(plotattributes, :yformatter)
     end
 
     # handle grid args common to all axes